Remove ambiguous message on vim bootstrap
On first run, when the vimrc doesn't exist, if you tried to remove it,
this would produce a 'file does not exist' message, which left it
ambiguous as to whether or not the bootstrapping had failed. The seeming
duplication of the link command is also necessary for that same reason,
because it produces console output that leaves it uncertain as to
whether or not bootstrapping succeeded.