Remove the solarized colour scheme for now
For some reason that I can't figure out at present, the colour scheme is
only actually being applied to the status bar at the bottom, rather than
the whole window. It seems that gnome-terminal is somehow overriding it.
After a lot of googling, I couldn't find anything, so will come back to
it later.