Python script to use vim as the svn difftool
Note that this commit contains a definition of a tunnel that refers to
something in work. While I'd prefer not to commit this, if I want to be
able to symlink the config file, at the moment I don't know of any
other easy way of not putting this in. It doesn't really compromise
anything anyway, as the location is only accessible from within my work
VPN.