Change modifier to the windows key
It turns out using control for this is a really bad idea, as it
interferes with so many other control based shortcuts. I just need to
unlearn my habit of using control (or caps lock mapped to control) to
navigate windows (the setup I had on tmux). That shouldn't be too hard.