Elaborate to backup and restore images
Also convert gifs to jpeg. Some of the bandcamp releases are actually
being put out with gifs. Tried to do these with find and so on, but
to get access to the file names, it turned out it was way easier to
just use a loop.

For some reason, even if there are no gifs in the directory, that loop
body still executes. This is really annoying, as it gives the
impression that something went wrong with the script, but it's
perfectly normal that there might be no gifs in the directory.