Find and replace with fzf
The default 'alt-a' binding doesn't seem to work for adding everything
to the quickfix list, so I've explicitly bound it to 'ctrl-a'. There is
also a little function I found for removing items from the quickfix list
with 'dd'. Finally, <leader>ag will search for the word under cursor or
highlighted text.