Remove Old Configurations
Loads of stuff in here was years old and really not being used any more.
If I ever need any of it I can recover it from the diffs here.

The repo is now whittled down to everything I'm actually using now.

All the bootstrap scripts are also removed because I'm going to be
switching to using GNU Stow.

The AMU scripts were moved to the AMU repo (not sure why I didn't do
that in the first place).