Use GNU Stow for Bootstrap
The config files are now moved to the expected locations relative to my
home directory, ready to be used by Stow.

I thought you would be able to run `stow . --target=/home/chris` and it
would go into all the top level folders here, then symlink from there,
but that's apparently not the way it works, so I'm just using a Makefile
to create all the links.