Update default use of bat and use it for git diff
By default, bat will now only use syntax highlighting. I was also using it with the `changes` style
to only show changes from Git, but this is kind of annoying for copying and pasting, plus, when
there are no changes, it still prints out a blank column, which is kind of annoying for copying and
pasting too.

There's been a Fish alias created called 'gdiff' that pipes `git diff` through bat. In this case,
bat is used with the `full` style to get all the file headers and the line numbers. There's also an
alias for `cat` to go straight to `bat`.